03-react-agent/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langchain-anthropic==0.3.0
langgraph==0.2.55
